# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
    "/authenticate": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "body",
            "description": "Code for authentication.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "since",
            "description": "A user ID. Only return users with an ID greater than this ID..",
            "type": "string",
            "in": "query"
          },
          {
            "name": "per_page",
            "description": "Results per page (max 100).",
            "type": "string",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/{id}/repos": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "description": "Page number of the results to fetch.",
            "type": "string",
            "in": "query"
          },
          {
            "name": "per_page",
            "description": "Results per page (max 100).",
            "type": "string",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {}
}